+++ example/android/app/src/main/kotlin/com/pragtech/json_form_engine_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.pragtech.json_form_engine_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()